classifier: add --minclassfreq option that provides a floor for classifying variants
This optionally creates an unknown (UKN) class of variants where there
is not strong support for any variant classification. Uncertain
classifications, where the supports look like `WP=0.326,0.32,0.354`,
motivated this. Instead of trying to classify these, it instead will put
them into an unknown category to reflect this lack of certainty.